funcionalidad /get y /post/newUser con mapperStruct.
