+++ pom.xml
@@ -43,6 +43,17 @@
               <artifactId>mapstruct-processor</artifactId>
               <version>1.4.1.Final</version>
             </path>
+            <path>
+              <groupId>org.projectlombok</groupId>
+              <artifactId>lombok</artifactId>
+              <version>${lombok.version}</version>
+            </path>
+            <path>
+              <groupId>org.projectlombok</groupId>
+              <artifactId>lombok-mapstruct-binding</artifactId>
+              <version>0.2.0</version>
+            </path>
+
           </annotationProcessorPaths>
         </configuration>
       </plugin>
