+++ src/main/java/org/drg/springcourt/models/UserEntity.java
@@ -2,16 +2,16 @@ package org.drg.springcourt.models;
 
 import lombok.*;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import javax.persistence.*;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
+import java.util.List;
 
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
+@Getter
+@Setter
 @ToString
 @Entity(name = "users")
 public class UserEntity {
@@ -36,55 +36,8 @@ public class UserEntity {
     @NotBlank(message = "Password is mandatory")
     private String password;
 
- /*   @OneToMany(mappedBy = "user")
-    private List<AddressEntity> address;*/
+    @OneToMany(fetch = FetchType.LAZY, mappedBy = "user")
+    private List<AddressEntity> address;
 
 
-  public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getSurname() {
-        return surname;
-    }
-
-    public void setSurname(String surname) {
-        this.surname = surname;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getMail() {
-        return mail;
-    }
-
-    public void setMail(String mail) {
-        this.mail = mail;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-   }
 }
