+++ src/main/resources/application-dev.properties
@@ -1,6 +1,7 @@
 # spring traces
 logging.level.org.springframework.web=DEBUG
 
+
 # open-api
 springdoc.api-docs.path=/api-docs
 springdoc.swagger-ui.path=/springbooks-manager-api.html
